pyproject.toml
[project]
name = "terrazip-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
ai = [
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
]
x402 = [
    "eth-account>=0.13.7",
    "web3>=7.14.0",
]
